Added the DecksList screen with create and study buttons. Also created the study window, which for now is a place holder until we can get the swipe action down.
