+++ app/src/main/res/layout/activity_courses.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?android:attr/colorBackground"
-    tools:context="com.example.terin.asu_flashcardapp.StudyWindow">
+    tools:context="com.example.terin.asu_flashcardapp.CourseList">
 
     <ListView
         android:id="@+id/coursesListView"
