+++ app/src/main/res/layout/fragment_create_input.xml
@@ -43,7 +43,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
-        app:layout_constraintGuide_begin="175dp" />
+        app:layout_constraintGuide_begin="120dp" />
 
     <Button
         android:id="@+id/createButton"
